database now using new data in orig data files
also added comment about what needs to be implemented for the future when reading from the db files. There's a temporary fix for this we can do where we just edit the data using the gui and that automatically packages a multiple word string as 1 element in a list. Meaning it allows us to save multiple words for a thing.